fixed minor errors
https://app.asana.com/0/161555263951748/169698496837359

+++ public_html/php/apis/truck/index.php
@@ -61,9 +61,9 @@ try {
 			if($truck !== null) {
 				$reply->data = $truck;
 			}
-		}else{
+		} else {
 			$trucks = Truck::getAllTrucks($pdo);
-			if($trucks !== null){
+			if($trucks !== null) {
 				$reply->data = $trucks;
 			}
 		}
@@ -72,55 +72,59 @@ try {
 	} elseif((empty($_SESSION["profile"]) === false) && (($_SESSION["profile"]->getProfileId()) === $id) && (($_SESSION["profile"]->getProfileType()) === "a") || (($_SESSION["profile"]->getProfileType())) === "o") {
 
 		if($method === "PUT" || $method === "POST") {
-		verifyXsrf();
-		$requestContent = file_get_contents("php://input");
-		$requestObject = json_decode($requestContent);
+			verifyXsrf();
+			$requestContent = file_get_contents("php://input");
+			$requestObject = json_decode($requestContent);
 
 
-		//Do i even need to create these???
-		$profile = Profile::getProfileByProfileId($pdo, $id);
-		$company = Company::getCompanyByCompanyId($pdo, $id);
+			//Do i even need to create these???
+			$profile = Profile::getProfileByProfileId($pdo, $id);
+			$company = Company::getCompanyByCompanyId($pdo, $id);
 
-		if($requestObject->profileType !== 'a' || 'o') {
-			throw(new \InvalidArgumentException("profile type must be admin('a') or owner('o') in order to make changes to a truck"));
+			if($requestObject->profileType !== 'a' || 'o') {
+				throw(new \InvalidArgumentException("profile type must be admin('a') or owner('o') in order to make changes to a truck"));
+			}
+
+			//make sure the truck foreign key is available (required field)
+			if(empty($requestObject->truckCompanyId) === true) {
+				throw(new \InvalidArgumentException("No foreign Id for truck", 405));
+			}
+
+			if($method === "POST") {
+				//create a new truck and insert it into the database
+				$truck = new Truck(null, $requestObject->companyId);
+				$truck->insert($pdo);
+
+				//update reply
+				$reply->message = "Truck created A-OK";
+			}
+
+			//delete section here
+		} elseif((empty($_SESSION["profile"]) === false) && (($_SESSION["profile"]->getProfileId()) === $id) && (($_SESSION["profile"]->getProfileType()) === "a") || (($_SESSION["profile"]->getProfileType())) === "o") {
+
+			if($method === "DELETE") {
+				verifyXsrf();
+				//retrieve the company to be deleted
+				$company = Company::getCompanyByCompanyId($pdo, $id);
+
+				//check if empty
+				if($company === null) {
+					throw(new RuntimeException("The company does not exist", 404));
+				}
+
+				//delete the company
+				$company->delete($pdo);
+
+				//update the reply
+				$reply->message = "Company deleted A-OK";
+			} else {
+				throw (new InvalidArgumentException("Invalid HTTP method request"));
+			}
+
 		}
-
-		//make sure the truck foreign key is available (required field)
-		if(empty($requestObject->truckCompanyId) === true) {
-			throw(new \InvalidArgumentException("No foreign Id for truck", 405));
-		}
-
-		if($method === "POST") {
-			//craete a new truck and insert it into the database
-			$truck = new Truck(null, $requestObject->companyId);
-			$truck->insert($pdo);
-
-			//update reply
-			$reply->message = "Truck created A-OK";
-		}
-
-		//delete section here
-	} elseif($method === "DELETE") {
-		verifyXsrf();
-		//retrieve the company to be deleted
-		$company = Company::getCompanyByCompanyId($pdo, $id);
-
-		//check if empty
-		if($company === null) {
-			throw(new RuntimeException("The company does not exist", 404));
-		}
-
-		//delete the company
-		$company->delete($pdo);
-
-		//update the reply
-		$reply->message = "Company deleted A-OK";
-	} else {
-		throw (new InvalidArgumentException("Invalid HTTP method request"));
 	}
-
 	//end of try, begging of catches
-}catch(Exception $exception) {
+} catch(Exception $exception) {
 	$reply->status = $exception->getCode();
 	$reply->message = $exception->getMessage();
 } catch(TypeError $typeError) {
